# backend/docs/swagger.yaml
basePath: /api
definitions:
...
  dto.CourseDetailAlumnoResponse:
    properties:
      creadorNombre:
        type: string
      descripcion:
        type: string
      enrolled:
        description: always true
        type: boolean
      id:
        type: string
      materiales:
        items:
          $ref: '#/definitions/dto.MaterialResponse'
        type: array
      secciones:
        items:
          $ref: '#/definitions/dto.SectionWithVideosResponse'
        type: array
      titulo:
        type: string
    type: object
...
  dto.EnrollmentResponse:
    properties:
      courseId:
        type: string
      enrolled:
        description: always true on 200
        type: boolean
    type: object
...
  dto.MaterialResponse:
    properties:
      createdAt:
        type: string
      id:
        type: string
      mimeType:
        type: string
      nombre:
        description: mapped from MaterialModel.Titulo
        type: string
      tamanoBytes:
        type: integer
    type: object
  dto.MyCourseItem:
    properties:
      completado:
        type: boolean
      courseId:
        type: string
      creadorNombre:
        type: string
      inscritoEn:
        type: string
      titulo:
        type: string
    type: object
...
  dto.SectionWithVideosResponse:
    properties:
      courseId:
        type: string
      createdAt:
        type: string
      id:
        type: string
      orden:
        type: integer
      titulo:
        type: string
      videos:
        items:
          $ref: '#/definitions/dto.VideoResponse'
        type: array
    type: object
...
  dto.VideoResponse:
    properties:
      createdAt:
        type: string
      duracionS:
        type: integer
      id:
        type: string
      orden:
        type: integer
      proveedor:
        type: string
      sectionId:
        type: string
      titulo:
        type: string
      url:
        type: string
    type: object
...
  httperr.Error:
    properties:
      code:
        type: string
      details:
        additionalProperties:
          type: string
        type: object
      message:
        type: string
    type: object
host: localhost:3000
info:
  contact: {}
  description: Plataforma interna de formación en video — LMS corporativo
  title: SkillMaker API
  version: "1.0"
paths:
...
  /catalog:
    get:
      description: Retorna una página de cursos con estado='aprobado'. Soporta ?page,
        ?size, ?q (ILIKE titulo).
      parameters:
      - description: Página (default 1)
        in: query
        name: page
        type: integer
      - description: Tamaño de página (max 100, default 20)
        in: query
        name: size
        type: integer
      - description: Filtro ILIKE en titulo
        in: query
        name: q
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Página de CatalogCourseCard (items, page, size, total, totalPages)
          schema:
            type: object
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/httperr.Error'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/httperr.Error'
      security:
      - BearerAuth: []
      summary: Lista cursos aprobados (catálogo público para alumnos)
      tags:
      - catalog
  /catalog/{id}:
    get:
      description: Retorna preview (enrolled=false) o árbol completo (enrolled=true)
        según inscripción.
      parameters:
      - description: UUID del curso
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: enrolled — árbol completo
          schema:
            $ref: '#/definitions/dto.CourseDetailAlumnoResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/httperr.Error'
        "404":
          description: curso no encontrado o no aprobado
          schema:
            $ref: '#/definitions/httperr.Error'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/httperr.Error'
      security:
      - BearerAuth: []
      summary: Detalle de un curso aprobado (alumno)
      tags:
      - catalog
  /catalog/{id}/enroll:
    post:
      description: 'Crea una fila de enrollment. Idempotente: segunda llamada devuelve
        200 sin duplicar.'
      parameters:
      - description: UUID del curso
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.EnrollmentResponse'
        "401":
          description: JWT requerido
          schema:
            $ref: '#/definitions/httperr.Error'
        "404":
          description: curso no encontrado o no aprobado
          schema:
            $ref: '#/definitions/httperr.Error'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/httperr.Error'
      security:
      - BearerAuth: []
      summary: Inscribe al alumno en un curso aprobado (idempotente)
      tags:
      - catalog
...
  /users/me/courses:
    get:
      description: Retorna los enrollments del caller con titulo, creadorNombre, completado,
        inscritoEn.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/dto.MyCourseItem'
            type: array
        "401":
          description: JWT requerido
          schema:
            $ref: '#/definitions/httperr.Error'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/httperr.Error'
      security:
      - BearerAuth: []
      summary: Lista los cursos en los que está inscrito el alumno autenticado
      tags:
      - catalog
...
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"